[fixcore][feat] Migrate existing resoto data and configs
